requirements.txt
torch
torchvision
torchaudio
transformers
open_clip_torch
datasets
accelerate
pycocoevalcap
nltk
scst